+++ packages/workbox-strategies/src/CacheOnly.ts
@@ -6,28 +6,29 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {assert} from 'workbox-core/_private/assert.mjs';
-import {cacheNames} from 'workbox-core/_private/cacheNames.mjs';
-import {fetchWrapper} from 'workbox-core/_private/fetchWrapper.mjs';
-import {logger} from 'workbox-core/_private/logger.mjs';
-import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
+import {assert} from 'workbox-core/_private/assert.js';
+import {cacheNames} from 'workbox-core/_private/cacheNames.js';
+import {cacheWrapper} from 'workbox-core/_private/cacheWrapper.js';
+import {logger} from 'workbox-core/_private/logger.js';
+import {WorkboxError} from 'workbox-core/_private/WorkboxError.js';
 
-import {messages} from './utils/messages.mjs';
-import './_version.mjs';
+import {messages} from './utils/messages.js';
+import './_version.js';
+
 
 /**
  * An implementation of a
- * [network-only]{@link https://developers.google.com/web/fundamentals/instant-and-offline/offline-cookbook/#network-only}
+ * [cache-only]{@link https://developers.google.com/web/fundamentals/instant-and-offline/offline-cookbook/#cache-only}
  * request strategy.
  *
  * This class is useful if you want to take advantage of any
  * [Workbox plugins]{@link https://developers.google.com/web/tools/workbox/guides/using-plugins}.
  *
- * If the network request fails, this will throw a `WorkboxError` exception.
+ * If there is no cache match, this will throw a `WorkboxError` exception.
  *
  * @memberof workbox.strategies
  */
-class NetworkOnly {
+class CacheOnly {
   /**
    * @param {Object} options
    * @param {string} options.cacheName Cache name to store and retrieve
@@ -35,14 +36,12 @@ class NetworkOnly {
    * [workbox-core]{@link workbox.core.cacheNames}.
    * @param {Array<Object>} options.plugins [Plugins]{@link https://developers.google.com/web/tools/workbox/guides/using-plugins}
    * to use in conjunction with this caching strategy.
-   * @param {Object} options.fetchOptions Values passed along to the
-   * [`init`](https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/fetch#Parameters)
-   * of all fetch() requests made by this strategy.
+   * @param {Object} options.matchOptions [`CacheQueryOptions`](https://w3c.github.io/ServiceWorker/#dictdef-cachequeryoptions)
    */
   constructor(options = {}) {
     this._cacheName = cacheNames.getRuntimeName(options.cacheName);
     this._plugins = options.plugins || [];
-    this._fetchOptions = options.fetchOptions || null;
+    this._matchOptions = options.matchOptions || null;
   }
 
   /**
@@ -85,42 +84,38 @@ class NetworkOnly {
     if (process.env.NODE_ENV !== 'production') {
       assert.isInstance(request, Request, {
         moduleName: 'workbox-strategies',
-        className: 'NetworkOnly',
-        funcName: 'handle',
+        className: 'CacheOnly',
+        funcName: 'makeRequest',
         paramName: 'request',
       });
     }
 
-    let error;
-    let response;
-    try {
-      response = await fetchWrapper.fetch({
-        request,
-        event,
-        fetchOptions: this._fetchOptions,
-        plugins: this._plugins,
-      });
-    } catch (err) {
-      error = err;
-    }
+    const response = await cacheWrapper.match({
+      cacheName: this._cacheName,
+      request,
+      event,
+      matchOptions: this._matchOptions,
+      plugins: this._plugins,
+    });
 
     if (process.env.NODE_ENV !== 'production') {
       logger.groupCollapsed(
-          messages.strategyStart('NetworkOnly', request));
+          messages.strategyStart('CacheOnly', request));
       if (response) {
-        logger.log(`Got response from network.`);
+        logger.log(`Found a cached response in the '${this._cacheName}'` +
+          ` cache.`);
+        messages.printFinalResponse(response);
       } else {
-        logger.log(`Unable to get a response from the network.`);
+        logger.log(`No response found in the '${this._cacheName}' cache.`);
       }
-      messages.printFinalResponse(response);
       logger.groupEnd();
     }
 
     if (!response) {
-      throw new WorkboxError('no-response', {url: request.url, error});
+      throw new WorkboxError('no-response', {url: request.url});
     }
     return response;
   }
 }
 
-export {NetworkOnly};
+export {CacheOnly};
