+++ packages/workbox-strategies/src/NetworkOnly.ts
@@ -6,29 +6,28 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {assert} from 'workbox-core/_private/assert.mjs';
-import {cacheNames} from 'workbox-core/_private/cacheNames.mjs';
-import {cacheWrapper} from 'workbox-core/_private/cacheWrapper.mjs';
-import {logger} from 'workbox-core/_private/logger.mjs';
-import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
+import {assert} from 'workbox-core/_private/assert.js';
+import {cacheNames} from 'workbox-core/_private/cacheNames.js';
+import {fetchWrapper} from 'workbox-core/_private/fetchWrapper.js';
+import {logger} from 'workbox-core/_private/logger.js';
+import {WorkboxError} from 'workbox-core/_private/WorkboxError.js';
 
-import {messages} from './utils/messages.mjs';
-import './_version.mjs';
-
+import {messages} from './utils/messages.js';
+import './_version.js';
 
 /**
  * An implementation of a
- * [cache-only]{@link https://developers.google.com/web/fundamentals/instant-and-offline/offline-cookbook/#cache-only}
+ * [network-only]{@link https://developers.google.com/web/fundamentals/instant-and-offline/offline-cookbook/#network-only}
  * request strategy.
  *
  * This class is useful if you want to take advantage of any
  * [Workbox plugins]{@link https://developers.google.com/web/tools/workbox/guides/using-plugins}.
  *
- * If there is no cache match, this will throw a `WorkboxError` exception.
+ * If the network request fails, this will throw a `WorkboxError` exception.
  *
  * @memberof workbox.strategies
  */
-class CacheOnly {
+class NetworkOnly {
   /**
    * @param {Object} options
    * @param {string} options.cacheName Cache name to store and retrieve
@@ -36,12 +35,14 @@ class CacheOnly {
    * [workbox-core]{@link workbox.core.cacheNames}.
    * @param {Array<Object>} options.plugins [Plugins]{@link https://developers.google.com/web/tools/workbox/guides/using-plugins}
    * to use in conjunction with this caching strategy.
-   * @param {Object} options.matchOptions [`CacheQueryOptions`](https://w3c.github.io/ServiceWorker/#dictdef-cachequeryoptions)
+   * @param {Object} options.fetchOptions Values passed along to the
+   * [`init`](https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/fetch#Parameters)
+   * of all fetch() requests made by this strategy.
    */
   constructor(options = {}) {
     this._cacheName = cacheNames.getRuntimeName(options.cacheName);
     this._plugins = options.plugins || [];
-    this._matchOptions = options.matchOptions || null;
+    this._fetchOptions = options.fetchOptions || null;
   }
 
   /**
@@ -84,38 +85,42 @@ class CacheOnly {
     if (process.env.NODE_ENV !== 'production') {
       assert.isInstance(request, Request, {
         moduleName: 'workbox-strategies',
-        className: 'CacheOnly',
-        funcName: 'makeRequest',
+        className: 'NetworkOnly',
+        funcName: 'handle',
         paramName: 'request',
       });
     }
 
-    const response = await cacheWrapper.match({
-      cacheName: this._cacheName,
-      request,
-      event,
-      matchOptions: this._matchOptions,
-      plugins: this._plugins,
-    });
+    let error;
+    let response;
+    try {
+      response = await fetchWrapper.fetch({
+        request,
+        event,
+        fetchOptions: this._fetchOptions,
+        plugins: this._plugins,
+      });
+    } catch (err) {
+      error = err;
+    }
 
     if (process.env.NODE_ENV !== 'production') {
       logger.groupCollapsed(
-          messages.strategyStart('CacheOnly', request));
+          messages.strategyStart('NetworkOnly', request));
       if (response) {
-        logger.log(`Found a cached response in the '${this._cacheName}'` +
-          ` cache.`);
-        messages.printFinalResponse(response);
+        logger.log(`Got response from network.`);
       } else {
-        logger.log(`No response found in the '${this._cacheName}' cache.`);
+        logger.log(`Unable to get a response from the network.`);
       }
+      messages.printFinalResponse(response);
       logger.groupEnd();
     }
 
     if (!response) {
-      throw new WorkboxError('no-response', {url: request.url});
+      throw new WorkboxError('no-response', {url: request.url, error});
     }
     return response;
   }
 }
 
-export {CacheOnly};
+export {NetworkOnly};
