+++ packages/workbox-strategies/src/plugins/cacheOkAndOpaquePlugin.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import '../_version.mjs';
+import '../_version.js';
 
 export const cacheOkAndOpaquePlugin = {
   /**
